+++ example/myshop/models/i18n_smartcard/__init__.py
@@ -32,7 +32,7 @@ class SmartCard(CMSPageReferenceMixin, TranslatableModelMixin, BaseProduct):
     product_name = models.CharField(max_length=255, verbose_name=_("Product Name"))
     slug = models.SlugField(verbose_name=_("Slug"))
     unit_price = MoneyField(_("Unit price"), decimal_places=3,
-        help_text=_("Net price for this product"))
+                            help_text=_("Net price for this product"))
     caption = TranslatedField()
     description = TranslatedField()
 
@@ -45,12 +45,12 @@ class SmartCard(CMSPageReferenceMixin, TranslatableModelMixin, BaseProduct):
     speed = models.CharField(_("Transfer Speed"), choices=SPEED, max_length=8)
     product_code = models.CharField(_("Product code"), max_length=255, unique=True)
     storage = models.PositiveIntegerField(_("Storage Capacity"),
-        help_text=_("Storage capacity in GB"))
+                                          help_text=_("Storage capacity in GB"))
 
     # controlling the catalog
     order = models.PositiveIntegerField(verbose_name=_("Sort by"), db_index=True)
     cms_pages = models.ManyToManyField('cms.Page', through=ProductPage,
-        help_text=_("Choose list view this product shall appear on."))
+                                       help_text=_("Choose list view this product shall appear on."))
     images = models.ManyToManyField('filer.Image', through=ProductImage)
 
     objects = ProductManager()
@@ -77,14 +77,13 @@ class SmartCard(CMSPageReferenceMixin, TranslatableModelMixin, BaseProduct):
 
 
 class SmartCardTranslation(TranslatedFieldsModel):
-    master = models.ForeignKey(SmartCard, related_name='translations',
-        null=True)
+    master = models.ForeignKey(SmartCard, related_name='translations', null=True)
     caption = HTMLField(verbose_name=_("Caption"),
-        configuration='CKEDITOR_SETTINGS_CAPTION',
-        help_text=_("Short description used in the catalog's list view of products."))
+                        configuration='CKEDITOR_SETTINGS_CAPTION',
+                        help_text=_("Short description used in the catalog's list view of products."))
     description = HTMLField(verbose_name=_("Description"),
-        configuration='CKEDITOR_SETTINGS_DESCRIPTION',
-        help_text=_("Description for the list view of products."))
+                            configuration='CKEDITOR_SETTINGS_DESCRIPTION',
+                            help_text=_("Description for the list view of products."))
 
     class Meta:
         unique_together = [('language_code', 'master')]