+++ example/tests/test_cart.py
@@ -2,13 +2,10 @@
 from __future__ import unicode_literals
 
 from django.conf import settings
-from django.contrib.auth import get_user_model
 from django.core.urlresolvers import reverse
 from rest_framework.test import APIClient, APIRequestFactory
 from cms.api import add_plugin, create_page
 from shop.models.cart import CartModel
-from shop.models.customer import CustomerState
-from shop.models.defaults.customer import Customer
 from shop.cascade.cart import ShopCartPlugin
 from myshop.models.polymorphic.smartcard import SmartCard
 from .test_shop import ShopTestCase
@@ -20,7 +17,7 @@ class CartTest(ShopTestCase):
 
         # create a page populated with Cascade elements used for checkout
         self.cart_page = create_page("Cart", 'INHERIT', 'en', parent=self.home_page,
-                                    published=True, in_navigation=True)
+                                     published=True, in_navigation=True)
 
     def add_editable_cart(self):
         """Add one GuestFormPlugin to the current page"""
@@ -112,30 +109,22 @@ class CartTest(ShopTestCase):
         self.assertEquals(len(response.data['items']), 0)
 
     def test_merge_items(self):
-        BART = {
-            'first_name': 'Bart',
-            'last_name': 'Simpson',
-            'email': 'bart@thesimpsons.com',
-            'password': 'trab',
-        }
-        bart = get_user_model().objects.create_user('bart', **BART)
-        bart = Customer.objects.create(user=bart, salutation='mr',
-                                       recognized=CustomerState.REGISTERED)
-        self.assertTrue(bart.is_registered())
-
         self.client.login(username='bart', password='trab')
         sessionid = self.client.session.session_key
 
-        # Bart adds a product
+        # Bart adds two products
         sdhc_4gb = SmartCard.objects.get(slug="sdhc-card-4gb")
         self.add_product2cart(sdhc_4gb)
+        xtr_sdhc_16gb = SmartCard.objects.get(slug="extreme-plus-sdhc-16gb")
+        self.add_product2cart(xtr_sdhc_16gb)
 
         # check that the added item is in the cart
         request = self.factory.get('/')
         self.middleware_process_request(request, sessionid)
         cart = CartModel.objects.get_from_request(request)
-        self.assertEquals(cart.items.count(), 1)
-        self.assertEquals(cart.items.first().product.product_code, sdhc_4gb.product_code)
+        self.assertEquals(cart.items.count(), 2)
+        self.assertEquals(cart.items.all()[0].product.product_code, sdhc_4gb.product_code)
+        self.assertEquals(cart.items.all()[1].product.product_code, xtr_sdhc_16gb.product_code)
 
         self.client.logout()
         self.assertNotEquals(sessionid, self.client.session.session_key)
@@ -147,9 +136,10 @@ class CartTest(ShopTestCase):
         with self.assertRaises(CartModel.DoesNotExist):
             CartModel.objects.get_from_request(request)
 
-        # add another item anonymously
-        xtr_sdhc_16gb = SmartCard.objects.get(slug="extreme-plus-sdhc-16gb")
+        # add two items anonymously
         self.add_product2cart(xtr_sdhc_16gb)
+        sdxc_pro_32gb = SmartCard.objects.get(slug="extreme-pro-micro-sdhc-32gb")
+        self.add_product2cart(sdxc_pro_32gb)
 
         # Bart logs in again
         login_url = reverse('shop:login')
@@ -157,9 +147,14 @@ class CartTest(ShopTestCase):
         response = self.client.post(login_url, credentials)
         self.assertEqual(response.status_code, 200)
 
-        # the cart now should contain both items
+        # the cart now should contain three items
         request = APIRequestFactory().request()
         self.middleware_process_request(request, self.client.session.session_key)
         cart = CartModel.objects.get_from_request(request)
-        self.assertEquals(cart.items.count(), 2)
-        self.assertEquals(cart.items.last().product.product_code, xtr_sdhc_16gb.product_code)
+        self.assertEquals(cart.items.count(), 3)
+        self.assertEquals(cart.items.all()[0].product.product_code, sdhc_4gb.product_code)
+        self.assertEquals(cart.items.all()[0].quantity, 1)
+        self.assertEquals(cart.items.all()[1].product.product_code, xtr_sdhc_16gb.product_code)
+        self.assertEquals(cart.items.all()[1].quantity, 2)
+        self.assertEquals(cart.items.all()[2].product.product_code, sdxc_pro_32gb.product_code)
+        self.assertEquals(cart.items.all()[2].quantity, 1)