+++ shop/apps.py
@@ -20,12 +20,6 @@ class ShopConfig(AppConfig):
 
         post_transition.connect(order_event_notification)
 
-        # Monkey patches for Django-1.8
-        if get_tuple_version()[:2] < (1, 9):
-            from django.utils import numberformat
-            from shop.patches import numberformat as patched_numberformat
-            numberformat.format = patched_numberformat.format
-
         # add JSONField to the map of customized serializers
         ModelSerializer.serializer_field_mapping[JSONField] = JSONSerializerField