+++ shop/cascade/catalog.py
@@ -51,7 +51,7 @@ plugin_pool.register_plugin(ShopAddToCartPlugin)
 
 class ProductGalleryForm(ModelForm):
     product = ProductSelectField(required=False, label=_("Related Product"),
-        help_text=_("Chose related product"))
+        help_text=_("Choose related product"))
 
     class Meta:
         exclude = ('glossary',)