+++ docs/changelog.rst
@@ -4,6 +4,16 @@
 Changelog for django-SHOP
 =========================
 
+0.9.4
+=====
+* In the backend, OrderAdmin and OrderItemAdmin may render the extra dicts from the model using
+  a special template.
+* Drop support for Django-1.8.
+* If an anonymous customer logs in, his current cart is merged with a cart, which has previously
+  been created. This has been adopted to re-use the method Product.is_in_cart()
+  in and finds it's Merge the contents of the other cart into this one, afterwards delete it.
+* Added method ``post_process_cart_item`` to the Cart Modifiers.
+
 0.9.3
 =====
 * Added template context processor :func:`shop.context_processors.ng_model_options` to add the